ai-product-video/backend/requirements.txt
fastapi
uvicorn
python-multipart
pillow
moviepy